authbox::pki

Module detail

Compatibility helpers for the C++ authbox::pik::detail namespace.

The C++ headers reopen authbox::pik::detail from several PKI headers. This module starts that shared namespace with the PEM/base64 helpers from xtra/authbox/include/pki/pem.hpp, pki/asn1_utils.hpp, pki/oid_registry.hpp, and pki/distinguished_name.hpp while keeping the real surfaces in their mirrored Rust modules.

Re-exports

pub use super::crl::CrlDerCursor;

pub use super::crl::crl_copy_span;

pub use super::crl::crl_parse_algorithm_identifier;

pub use super::crl::crl_parse_name;

pub use super::crl::crl_parse_time;

pub use super::crl::identify_crl_entry_extension;

pub use super::crl::identify_crl_extension;

pub use super::crl::parse_crl_entry_extensions;

pub use super::crl::parse_crl_extensions;

pub use super::crl::parse_reason_code;

pub use super::distinguished_name::ParsedDn;

pub use super::distinguished_name::attribute_from_oid;

pub use super::distinguished_name::attribute_from_string;

pub use super::distinguished_name::dn_trim;

pub use super::distinguished_name::encode_directory_string;

pub use super::distinguished_name::encode_name;

pub use super::distinguished_name::is_printable_string;

pub use super::distinguished_name::oid_from_attribute;

pub use super::distinguished_name::parse_dn_string;

pub use super::parser::ParseResult;

pub use super::parser::ValidityRange;

pub use super::parser::copy_bytes;

pub use super::parser::make_error;

pub use super::parser::parse_algorithm_identifier_full;

pub use super::parser::parse_certificate;

pub use super::parser::parse_extension_sequence;

pub use super::parser::parse_extensions_full;

pub use super::parser::parse_name_full;

pub use super::parser::parse_subject_public_key_info_full;

pub use super::parser::parse_time_choice as parse_time_choice_full;

pub use super::parser::parse_validity;

pub use super::parser::parse_version;

Structs

DnSpanCursor

Enums

KeylockSignatureAlgorithm

Constants

K_BASE64_ALPHABET

K_BEGIN_MARKER

K_CURVE_OIDS

K_END_MARKER

K_EXTENSION_OIDS

K_HASH_ALGORITHMS

K_MAX_LENGTH_OCTETS

K_OID_AUTHORITY_INFO_ACCESS

K_OID_AUTHORITY_KEY_ID

K_OID_BASIC_CONSTRAINTS

K_OID_BLAKE2B

K_OID_CERTIFICATE_POLICIES

K_OID_CRL_DISTRIBUTION_POINTS

K_OID_ECDSA_SHA256

K_OID_ECDSA_SHA384

K_OID_ECDSA_SHA512

K_OID_ED448

K_OID_ED25519

K_OID_EXTENDED_KEY_USAGE

K_OID_INHIBIT_ANY_POLICY

K_OID_ISSUER_ALT_NAME

K_OID_KEY_USAGE

K_OID_NAME_CONSTRAINTS

K_OID_POLICY_CONSTRAINTS

K_OID_POLICY_MAPPINGS

K_OID_RSA_PSS

K_OID_SECP256K1

K_OID_SECP256R1

K_OID_SECP384R1

K_OID_SECP521R1

K_OID_SHA256

K_OID_SHA512

K_OID_SHA256_WITH_RSA

K_OID_SHA384_WITH_RSA

K_OID_SHA512_WITH_RSA

K_OID_SUBJECT_ALT_NAME

K_OID_SUBJECT_KEY_ID

K_OID_X448

K_OID_X25519

K_SIGNATURE_ALGORITHMS

K_TRAILER

kBase64Alphabet

kBeginMarker

kCurveOids

kEndMarker

kExtensionOids

kHashAlgorithms

kMaxLengthOctets

kOidAuthorityInfoAccess

kOidAuthorityKeyId

kOidBasicConstraints

kOidBlake2b

kOidCertificatePolicies

kOidCrlDistributionPoints

kOidEcdsaSha256

kOidEcdsaSha384

kOidEcdsaSha512

kOidEd448

kOidEd25519

kOidExtendedKeyUsage

kOidInhibitAnyPolicy

kOidIssuerAltName

kOidKeyUsage

kOidNameConstraints

kOidPolicyConstraints

kOidPolicyMappings

kOidRsaPss

kOidSecp256k1

kOidSecp256r1

kOidSecp384r1

kOidSecp521r1

kOidSha256

kOidSha512

kOidSha256WithRsa

kOidSha384WithRsa

kOidSha512WithRsa

kOidSubjectAltName

kOidSubjectKeyId

kOidX448

kOidX25519

kSignatureAlgorithms

kTrailer

Functions

build_block

contains_pem_marker

decode_base64

decode_base64_char

encode_base64

equals_case_insensitive

error_result

get_extension_oid

get_signature_oid

ipv4_equal_bytes

is_digit

is_ipv4_literal

is_utctime

keylock_signature_algorithm

label_count

lookup_enum

make_certificate_from_context

make_random_serial

make_time_point

normalize_public_key_for_verify

normalize_signature_for_emit

normalize_signature_for_verify

parse_decimal

parse_extended_key_usage

parse_ipv4

parse_rsa_public_key_der

strip_integer_padding

strip_whitespace

to_gmtime

wildcard_match

Type Aliases

OidArray